fix #5 fix #6
