+++ egs_s2/LibriLight/local/preprocess.sh
@@ -123,7 +123,7 @@ if [ ${stage} -le 4 ] && [ ${stop_stage} -ge 4 ]; then
     echo "get_acoustic_token_librilight.py for ${sub_dataset} start!"
     for rank_id in {0..2}; do
         gpu_id=$((rank_id))
-        CUDA_VISIBLE_DEVICES=${gpu_id} python3 ${BIN_DIR}/get_acoustic_token_librilight.py \
+        OMP_NUM_THREADS=1 CUDA_VISIBLE_DEVICES=${gpu_id} python3 ${BIN_DIR}/get_acoustic_token_librilight.py \
             --data_dir=${data_dir} \
             --sub_dataset=${sub_dataset} \
             --dump_dir=${root_dir}/${dump_dir} \
@@ -148,7 +148,7 @@ if [ ${stage} -le 5 ] && [ ${stop_stage} -ge 5 ]; then
     echo "get_acoustic_token_librilight.py for ${sub_dataset} start!"
     for rank_id in {0..3}; do
         gpu_id=$((rank_id))
-        CUDA_VISIBLE_DEVICES=${gpu_id} python3 ${BIN_DIR}/get_acoustic_token_librilight.py \
+        OMP_NUM_THREADS=1 CUDA_VISIBLE_DEVICES=${gpu_id} python3 ${BIN_DIR}/get_acoustic_token_librilight.py \
             --data_dir=${data_dir} \
             --sub_dataset=${sub_dataset} \
             --dump_dir=${root_dir}/${dump_dir} \
@@ -173,7 +173,7 @@ if [ ${stage} -le 6 ] && [ ${stop_stage} -ge 6 ]; then
     echo "get_acoustic_token_librilight.py for ${sub_dataset} start!"
     for rank_id in {0..15}; do
         gpu_id=$((rank_id / 2))
-        CUDA_VISIBLE_DEVICES=${gpu_id} python3 ${BIN_DIR}/get_acoustic_token_librilight.py \
+        OMP_NUM_THREADS=1 CUDA_VISIBLE_DEVICES=${gpu_id} python3 ${BIN_DIR}/get_acoustic_token_librilight.py \
             --data_dir=${data_dir} \
             --sub_dataset=${sub_dataset} \
             --dump_dir=${root_dir}/${dump_dir} \
@@ -198,19 +198,19 @@ if [ ${stage} -le 7 ] && [ ${stop_stage} -ge 7 ]; then
     echo "get_acoustic_token_librilight.py for ${sub_dataset} start!"
     for rank_id in {0..3}; do
         gpu_id=$((rank_id))
-        CUDA_VISIBLE_DEVICES=${gpu_id} python3 ${BIN_DIR}/get_acoustic_token_librilight.py \
-        --data_dir=${data_dir} \
-        --sub_dataset=${sub_dataset} \
-        --dump_dir=${root_dir}/${dump_dir} \
-        --codec_name=hificodec \
-        --model_path=pretrained_model/hificodec/HiFi-Codec-16k-320d-large-universal \
-        --config_path=pretrained_model/hificodec/config_16k_320d.json \
-        --sr=16000 \
-        --num-cpu=24 \
-        --VAD_path=VAD/librilight_segment_dict.npy \
-        --nshard=4 \
-        --rank=${rank_id} &
-        eval pid${rank_id}="$!"
+        OMP_NUM_THREADS=1 CUDA_VISIBLE_DEVICES=${gpu_id} python3 ${BIN_DIR}/get_acoustic_token_librilight.py \
+            --data_dir=${data_dir} \
+            --sub_dataset=${sub_dataset} \
+            --dump_dir=${root_dir}/${dump_dir} \
+            --codec_name=hificodec \
+            --model_path=pretrained_model/hificodec/HiFi-Codec-16k-320d-large-universal \
+            --config_path=pretrained_model/hificodec/config_16k_320d.json \
+            --sr=16000 \
+            --num-cpu=24 \
+            --VAD_path=VAD/librilight_segment_dict.npy \
+            --nshard=4 \
+            --rank=${rank_id} &
+            eval pid${rank_id}="$!"
     done
     wait "$pid0" "$pid1" "$pid2" "$pid3"
     echo "get_acoustic_token_librilight.py for ${sub_dataset} done!"
